cache-http-client/requirements.txt
Flask==2.0.1
python-memcached==1.59